lib: lite_buttons: add define for minimum detection delay
Add define LITE_BUTTONS_DETECTION_DELAY_MIN_US, which holds
the minimum configurable button detection delay.

The reason for the minimum delay to be three times the
minimum timer value is because the timer needs to trigger
three times before the button is detected as pressed and
the user handler is called. The gpiote interrupt starts
the first timer. When it elapses, the button state transitions
from IDLE to PRESS_ARMED. The 2nd timer starts. When it elapses,
the state transitions from PRESS_ARMED to PRESS_DETECTED.
The 3rd timer starts. When it elapses, the state transitions from
PRESS_DETECTED to PRESSED and user_event() is invoked.

Add LITE_TIMER_US_TO_TICKS macro to convert from microseconds
to ticks. It is required in lite_buttons.

Update samples to use the new LITE_BUTTONS_DETECTION_DELAY_MIN_US
so that the configuration will be okay, even when the value of
CONFIG_SYS_CLOCK_TICKS_PER_SEC changes.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>