lib: lite_buttons: fix internal button states
The previous implementation would keep track of the pin
states using a bit field where each bit is assigned to
a pin based on its number. This representation brings
some problems in those cases where the pin number is
higher than the maximum size of `pin_active` (32).
For instance, if pin is number 13 of port 1, then the final
absolute pin number is ((1 << 5) + 13) which is 45. This
goes beyond the representable range of `pin_active` bit field.
To prevent this issue, the current implementation uses
instead the index of the pin within the configuration array
passed to the buttons library, this way there is always a
unique mapping 1:1 between a pin and its state in the buttons
library, while also being maximally compact in its representation
since, at most, the configuration array will have a maximum
number of configurable pins lower or equal than 32 for nRF52 SoC
and 24 for nRF54L SoC.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>