softdevice: let the NVIC state be part of the SoftDevice component
Let the global NVIC state be part of the SoftDevice component
(CONFIG_SOFTDEVICE) instead of the SoftDevice handler (CONFIG_NRF_SDH).

This is more correct, as it is possible to use nrf_nvic.h
without the SoftDevice handler.

It may also be helpful in tests or to swich more cleanly on
CONFIG_SOFTDEVICE instead of CONFIG_NRF_SDH when using NVIC functions.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>