bluetooth: services: bas: fix how battery level CCCD is added
Fix a bug in battery service where setting the 'can_notify'
configuration to false would cause the softdevice to return
invalid parameter when adding the battery level characteristic.

The reason was that char_md.p_cccd_md would always point to
the CCCD metadata structure. Even when 'can_notify' was set to
false and the CCCD metadata was left unpopulated.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>