nrf_sdh: let nrf_sdh_ble_enable() take the connection config tag
Let nrf_sdh_ble_enable() take the connection configuration tag
instead of the RAM start address. The library can figure out the RAM
start address by itself, there shouldn't be a need to pass it as
a parameter anymore.

Take instead the connection configuration tag, to let the SoftDevice
enable the BLE stack with a different configurations.

Remove the nrf_sdh_ble_default_cfg_set() call and use instead a Kconfig
to let the user choose which tag to use for automatic configuration.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>