bluetooth: services: nus: fix unsafe use of conn_handle as array idx
It is not safe to assume that connection handles will start at zero
and always stay below the number of max connections. Therefore,
connection handle cannot, and should not, be used to index into
arrays or lists.

Fix this by using the nrf_sdh_ble_idx_get function to return
an index that has been assigned to the connection handle
by the softdevice handler.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>