samples: fix processing of deferred logging when returning from main
If deferred logging is used (usually the sample default), logs will
have to be processed with LOG_PROCESS() before returning/exiting
from main. If this is not done, log messages will not be sent
over serial and not appear on terminal.

This fixes log messages being stuck in the logging buffer when
for instance an initialization function would fail, then print
some (important) message before exiting main.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>