boards: dts: Update memory layout for nRF54L
The softdevice targets for the nRF54L series had the
softdevice memory placed on the starting address of RAM
and the application RAM placed next.

This is problematic because the Cracen driver for these
devices assumes that it owns the top 64 bytes of memory
(starting from 0x2000000).

Rearrange the memory layout of these devices so that it
it places the softdevice memory is placed after the
application memory.

I also noticed that the partitions softdevice_static_ram
and softdevice_dynamic_ram had a gap between them.
I removed this gap since I cannot see why it is needed.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>