lib: bm_zms: fix writing when length is not a multiple of program unit
Fix writing when length is above the program unit and not a multiple
of the program unit. Program unit is 16.

Data was not offset correctly when writing the last few bytes.
Also, the temporary buffer used to pad the last few bytes went out of
scope before the data could be asynchronously written.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>