sdh: introduce observer priorities (high, lowest, etc.)
Introduce an allowed set of priorities for observers, with validity checks.
These are: HIGHEST, HIGH, USER, USER_LOW, LOWEST.

These could be extended ala SYS_INIT to include a priority within a level,
but it seemed a bit overkill for now.

Each library or piece of code in the SDK shall have an hardcoded priority,
because just changing the priority of one component would break the others.
In general, a component's observer must have a lower priority than the
observers in its dependencies.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>