sdh: ble: export nrf_sdh_ble_evt_tostr() and nrf_sdh_soc_evt_tostr()
Rework the functions in nrf_sdh_ble and nrf_sdh_soc that mapped BLE and
SOC events to their stringified version to make them public and align
their names. Instead of printing "unknown" if the event is not known,
default to printing its numerical value in hexadecimal.

That is also the default behavior if CONFIG_NRF_SDH_STR_TABLES is unset,
and the table is not compiled. This is better than toggling the
availability of the whole function, because we avoid conditional
compilation around it.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>