sdh: merge request/state observers
There was a distinction between observers that could stop the SoftDevice
state changes, and those that didn't. The former were called
"request observers" and the latter "state observers".

Since both observers are interested in the SoftDevice state,
this commit merges the two, keeping only "state observers",
for the purpose of simplifying things a bit.

Now, state observers can return non-zero to these events:
- NRF_SDH_STATE_EVT_ENABLE_PREPARE and
- NRF_SDH_STATE_EVT_DISABLE_PREPARE
to halt the state change.

The return value from the observer is ignored for other events.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>