sdh: remove nrf_sdh_request_continue()
This function shifts the responsibility of changing the state of the
SoftDevice to the observer that halted its state change. This means
that potentially all observers that halt the state change not only
become responsible of the process (which may not be ideal) but also
have to maintain some state and logic to restart that process.

Instead, let the component who requested the state change in the first
place retry the operation.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>