lib: peer_manager: remove sdk_macros
The verification macros in sdk_macros.h was copied from nRF5 SDK
to support quicker porting of the peer_manager and was never intended
to persist long in nRF Bare Metal option.
This removes the sdk_macros.h header file and substitutes the logic
where the VERIFY_xxx macros were used. This preserves the logic.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>