lib: peer_manager: id_manager: wrong wlist_addr_cnt input value
When enabling the debug checks (asserts), a value of zero for
wlist_addr_cnt triggers an assert in the peers_id_keys_get()
function. Correct it to the number of addresses that can be held
in the local addrs array passed to the peers_id_keys_get() function.

For context, on return from peers_id_keys_get(), the wlist_addr_cnt
will hold the number of addresses written to the addrs array.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>