.gitattributes: Unset diff attribute for pdf and hex files
The `GitDiffCheck` check complained about trailing whitespace
in the exported files from SoftDevice. The .pdf and .hex files
are build outputs and we do not care about the whitespace in them.

According to the git documentation, this is the correct way
to mark non-human-readable files:
> https://git-scm.com/docs/gitattributes#_marking_files_as_binary:
> ... because the content, while technically composed of text
> characters, is opaque to a human reader. ...

This makes `git diff --check` ignore the files.

Signed-off-by: Herman Berget <herman.berget@nordicsemi.no>