sdh: seed SoftDevice rand before enabling SD_EVT_IRQn
Upon enabling the SoftDevice events IRQ, the SoftDevice will request a
rand seed. When the SoftDevice events are dispatched by the scheduler,
it won't be possible to enable Bluetooth until that event has been
processed (in the main() loop).

To avoid this, we seed the SoftDevice before enabling the interrupt so
that no event is generated, and the application can complete the BLE
initialization before reaching the main() loop.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>