boards: bm_nrf54l15dk: fix grtc clock source selection
Explicitly setting the GRTC clock source in the board defconfig
causes build warning in the mcuboot configuration step.
LFXO is already the default and disabling the lfxo dts node
without changing this explicitly set value would lead to
an invalid configuration.

Remove the explicit GRTC clock source selection in defconfig
and give a new default to the GRTC clock source choice in case
the lfxo dts node is disabled.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>