subsys: storage: flash_map: Preserve firmware loader from himself
The firmware loader shall not be allowed to modify its own partition.
It was caught that it could erase itself.

This patch mitigates the issue by removal of the slot1_partition
mapping (it belongs to the firmware loader). MCUmgr couldn't erase
the partition as it is not mapped anymore.

ref.: NCSDK-36684

Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>