applications, samples: do not set sys attrs from connected event
Wait for the SoftDevice to send the BLE_GATTS_EVT_SYS_ATTR_MISSING
event before setting the GATTS system attributes with
sd_ble_gatts_sys_attr_set().

It is not illegal or wrong to set it on a connection event, but
it duplicates some code and it overcomplicates the samples a bit.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>