subsys: mgmt: mcumgr: img_mgmt: prevent self destruction of the loader
This is source code copy of:
https://github.com/zephyrproject-rtos/zephyr/pull/100689

The img_mgmt_mcumgr.c file can be switched back to sdk-zephyr once
origin patch will be aveilabel for referencing.

The FW loader reports and manages exactly two slots:
 - slot 0: this is the slot for the application code to update
 - slot 1: this is the slot, in which the FW loader is placed

The slot 1 is reported, so tools can fetch metadata about the
FW loader installed on the device.
Unfortunately, currently SMP-based FW loader allows to issue slot erase
command for the slot 1, effectively erasing the FW loader code that is
being executed.

This change correctly identifies the slot 1 as an active one, marking it
as used and blocking erase operation on that slot.

Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>