lib: bluetooth: ble_adv: fix allow list advertising without directed
With allow-list enabled, and directed and/or directed high duty
advertising disabled in Kconfig, the allow-list for use in fast and
slow advertising did not take effect when advertising was started with
mode directed or directed high duty.

This made it possible for non-paired/non-allow-listed devices to
connect after a paired/allow-listed device had disconnected and
advertising had been automatically restarted in the directed high
duty mode.

This fixes two internal checks in the ble_adv library that was
the source of the issue. These checks will now correctly
return the "capabilities" of the advertising mode to be
started so that the configuration will be correct.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>