lib: bluetooth: ble_adv: fix directed advertising set configuration
When configuring the advertisement set for direct advertisement
with sd_ble_gap_adv_set_configure(), NULL must be passed for the
advertising data argument because advertising data is not supported
in this mode. The sd_ble_gap_adv_set_configure() would return
NRF_ERROR_INVALID_PARAM in this case.

When doing directed advertisement, the address of the peer to direct
the advertisement towards needed to be specified in the advertisement
parameters when calling sd_ble_gap_adv_set_configure(). Set the address
in the advertisement parameters to the address supplied by the
application calling the ble_adv_peer_addr_reply() function in response
to the BLE_ADV_EVT_PEER_ADDR_REQUEST event.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>